# packages/types/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "TechRat API",
    "description": "Gamified tech learning platform. Authenticate with POST /api/v1/auth/login (bearer) or ?useCookies=true (web).",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5099/"
    }
  ],
  "paths": {
...
    "/api/v1/admin/questions": {
      "get": {
        "tags": [
          "Administration"
        ],
        "parameters": [
          {
            "name": "topic",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "difficulty",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Difficulty"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "active",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfAdminQuestionDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Administration"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminQuestionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminQuestionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/questions/{id}": {
      "get": {
        "tags": [
          "Administration"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminQuestionDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Administration"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminQuestionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminQuestionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/questions/{id}/deactivate": {
      "post": {
        "tags": [
          "Administration"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/admin/questions/{id}/activate": {
      "post": {
        "tags": [
          "Administration"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/admin/topics": {
      "post": {
        "tags": [
          "Administration"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminTopicInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/topics/{slug}": {
      "put": {
        "tags": [
          "Administration"
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminTopicInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/admin/topics/{slug}/subtopics": {
      "post": {
        "tags": [
          "Administration"
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminSubtopicInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/roadmaps": {
      "post": {
        "tags": [
          "Administration"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminRoadmapInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/roadmaps/{slug}": {
      "put": {
        "tags": [
          "Administration"
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminRoadmapInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/admin/roadmaps/{slug}/steps": {
      "post": {
        "tags": [
          "Administration"
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminStepInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/steps/{id}": {
      "put": {
        "tags": [
          "Administration"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminStepInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/admin/users": {
      "get": {
        "tags": [
          "Administration"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfAdminUserDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/roadmaps": {
      "get": {
        "tags": [
          "Roadmaps"
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoadmapSummaryDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/roadmaps/{slug}": {
      "get": {
        "tags": [
          "Roadmaps"
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoadmapDetailDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/topics/{slug}": {
      "get": {
        "tags": [
          "Topics"
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopicDetailDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminOptionDto": {
        "required": [
          "id",
          "text",
          "isCorrect",
          "displayOrder"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "text": {
            "type": "string"
          },
          "isCorrect": {
            "type": "boolean"
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "AdminOptionInput": {
        "required": [
          "text",
          "isCorrect"
        ],
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "isCorrect": {
            "type": "boolean"
          }
        }
      },
      "AdminQuestionDto": {
        "required": [
          "id",
          "externalKey",
          "topicSlug",
          "subtopicSlug",
          "difficulty",
          "questionType",
          "title",
          "questionText",
          "explanation",
          "referenceUrl",
          "xpReward",
          "isActive",
          "updatedAt",
          "options"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "externalKey": {
            "type": [
              "null",
              "string"
            ]
          },
          "topicSlug": {
            "type": "string"
          },
          "subtopicSlug": {
            "type": "string"
          },
          "difficulty": {
            "$ref": "#/components/schemas/Difficulty"
          },
          "questionType": {
            "$ref": "#/components/schemas/QuestionType"
          },
          "title": {
            "type": "string"
          },
          "questionText": {
            "type": "string"
          },
          "explanation": {
            "type": "string"
          },
          "referenceUrl": {
            "type": "string"
          },
          "xpReward": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminOptionDto"
            }
          }
        }
      },
      "AdminQuestionInput": {
        "required": [
          "topicSlug",
          "subtopicSlug",
          "difficulty",
          "title",
          "questionText",
          "explanation",
          "referenceUrl",
          "xpReward",
          "isActive",
          "options"
        ],
        "type": "object",
        "properties": {
          "topicSlug": {
            "type": "string"
          },
          "subtopicSlug": {
            "type": "string"
          },
          "difficulty": {
            "$ref": "#/components/schemas/Difficulty"
          },
          "title": {
            "type": "string"
          },
          "questionText": {
            "type": "string"
          },
          "explanation": {
            "type": "string"
          },
          "referenceUrl": {
            "type": "string"
          },
          "xpReward": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminOptionInput"
            }
          }
        }
      },
      "AdminRoadmapInput": {
        "required": [
          "slug",
          "name",
          "description",
          "category",
          "difficulty",
          "estimatedHours",
          "icon",
          "isPublished",
          "xpReward"
        ],
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "difficulty": {
            "$ref": "#/components/schemas/RoadmapDifficulty"
          },
          "estimatedHours": {
            "type": "integer",
            "format": "int32"
          },
          "icon": {
            "type": "string"
          },
          "isPublished": {
            "type": "boolean"
          },
          "xpReward": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "AdminStepInput": {
        "required": [
          "moduleId",
          "newModuleTitle",
          "title",
          "description",
          "difficulty",
          "estimatedMinutes",
          "topicSlug",
          "subtopicSlug",
          "minimumQuestions",
          "minimumAccuracy",
          "xpReward",
          "order"
        ],
        "type": "object",
        "properties": {
          "moduleId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "newModuleTitle": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "difficulty": {
            "$ref": "#/components/schemas/Difficulty"
          },
          "estimatedMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "topicSlug": {
            "type": "string"
          },
          "subtopicSlug": {
            "type": [
              "null",
              "string"
            ]
          },
          "minimumQuestions": {
            "type": "integer",
            "format": "int32"
          },
          "minimumAccuracy": {
            "type": "integer",
            "format": "int32"
          },
          "xpReward": {
            "type": "integer",
            "format": "int32"
          },
          "order": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          }
        }
      },
      "AdminSubtopicInput": {
        "required": [
          "slug",
          "name"
        ],
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "AdminTopicInput": {
        "required": [
          "slug",
          "name",
          "description",
          "category",
          "icon"
        ],
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "AdminUserDto": {
        "required": [
          "id",
          "username",
          "displayName",
          "email",
          "level",
          "xp",
          "questionsAnswered",
          "accuracy",
          "createdAt",
          "lastLoginAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "level": {
            "type": "integer",
            "format": "int32"
          },
          "xp": {
            "type": "integer",
            "format": "int64"
          },
          "questionsAnswered": {
            "type": "integer",
            "format": "int32"
          },
          "accuracy": {
            "type": "number",
            "format": "double"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastLoginAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "Difficulty": {
        "enum": [
          "Easy",
          "Medium",
          "Hard",
          "Expert"
        ]
      },
      "DifficultyBreakdownDto": {
        "required": [
          "difficulty",
          "answered",
          "correct",
          "accuracy"
        ],
        "type": "object",
        "properties": {
          "difficulty": {
            "type": "string"
          },
          "answered": {
            "type": "integer",
            "format": "int32"
          },
          "correct": {
            "type": "integer",
            "format": "int32"
          },
          "accuracy": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "IdResponse": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "PagedResultOfAdminQuestionDto": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminQuestionDto"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PagedResultOfAdminUserDto": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminUserDto"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "QuestionType": {
        "enum": [
          "MultipleChoice",
          "MultipleSelect",
          "TrueFalse",
          "CodingChallenge",
          "CodeOutput",
          "Debugging",
          "ArchitectureScenario"
        ]
      },
...
      "RoadmapDetailDto": {
        "required": [
          "summary",
          "prerequisites",
          "modules",
          "currentStepId"
        ],
        "type": "object",
        "properties": {
          "summary": {
            "$ref": "#/components/schemas/RoadmapSummaryDto"
          },
          "prerequisites": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoadmapPrerequisiteDto"
            }
          },
          "modules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoadmapModuleDto"
            }
          },
          "currentStepId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
      "RoadmapDifficulty": {
        "enum": [
          "Beginner",
          "Intermediate",
          "Advanced",
          "Expert"
        ]
      },
      "RoadmapModuleDto": {
        "required": [
          "id",
          "title",
          "order",
          "xpReward",
          "isCompleted",
          "steps"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "xpReward": {
            "type": "integer",
            "format": "int32"
          },
          "isCompleted": {
            "type": "boolean"
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoadmapStepDto"
            }
          }
        }
      },
      "RoadmapPrerequisiteDto": {
        "required": [
          "slug",
          "name",
          "minimumPercent",
          "currentPercent",
          "isMet"
        ],
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "minimumPercent": {
            "type": "integer",
            "format": "int32"
          },
          "currentPercent": {
            "type": "number",
            "format": "double"
          },
          "isMet": {
            "type": "boolean"
          }
        }
      },
...
      "RoadmapSummaryDto": {
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "category",
          "difficulty",
          "estimatedHours",
          "stepsCount",
          "modulesCount",
          "icon",
          "xpReward",
          "prerequisites",
          "progress"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "difficulty": {
            "type": "string"
          },
          "estimatedHours": {
            "type": "integer",
            "format": "int32"
          },
          "stepsCount": {
            "type": "integer",
            "format": "int32"
          },
          "modulesCount": {
            "type": "integer",
            "format": "int32"
          },
          "icon": {
            "type": "string"
          },
          "xpReward": {
            "type": "integer",
            "format": "int32"
          },
          "prerequisites": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "progress": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RoadmapUserStateDto"
              }
            ]
          }
        }
      },
      "RoadmapUserStateDto": {
        "required": [
          "isUnlocked",
          "isStarted",
          "isCompleted",
          "completedSteps",
          "percentComplete",
          "lastActivityAt"
        ],
        "type": "object",
        "properties": {
          "isUnlocked": {
            "type": "boolean"
          },
          "isStarted": {
            "type": "boolean"
          },
          "isCompleted": {
            "type": "boolean"
          },
          "completedSteps": {
            "type": "integer",
            "format": "int32"
          },
          "percentComplete": {
            "type": "number",
            "format": "double"
          },
          "lastActivityAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "SubtopicProgressDto": {
        "required": [
          "slug",
          "name",
          "questionCount",
          "answered",
          "correct",
          "accuracy"
        ],
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "questionCount": {
            "type": "integer",
            "format": "int32"
          },
          "answered": {
            "type": "integer",
            "format": "int32"
          },
          "correct": {
            "type": "integer",
            "format": "int32"
          },
          "accuracy": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "TopicDetailDto": {
        "required": [
          "topic",
          "progress",
          "subtopicProgress",
          "byDifficulty"
        ],
        "type": "object",
        "properties": {
          "topic": {
            "$ref": "#/components/schemas/TopicDto"
          },
          "progress": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TopicProgressDto"
              }
            ]
          },
          "subtopicProgress": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubtopicProgressDto"
            }
          },
          "byDifficulty": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DifficultyBreakdownDto"
            }
          }
        }
      },
      "TopicDto": {
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "category",
          "icon",
          "order",
          "questions",
          "subtopics"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "questions": {
            "$ref": "#/components/schemas/DifficultyCountsDto"
          },
          "subtopics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubtopicDto"
            }
          }
        }
      },
...
    }
  },
...
}